fix: Move version lib to internal & cleanup rp artifacts
rp stands for release-please
