+++ cmd/version.go
@@ -5,7 +5,7 @@ import (
 	"log"
 	"os"
 
-	"github.com/louislef299/aws-sso/pkg/version"
+	"github.com/louislef299/aws-sso/internal/version"
 	"github.com/spf13/cobra"
 )
